--- src/components/Card/card.module.css
@@ -1,32 +1,0 @@
-.cardContainer {
-
-background-color: rgba(231, 161, 115, 0.575);
-border: 1px solid grey;
-border-radius: 10px;
-border: none;
-padding: 20px;
-cursor: pointer;
--moz-osx-font-smoothing: grayscale;
-backface-visibility: hidden;
-transform: translateZ(0);
-transition: transform 0.25s ease-out;
-max-width: 20%;
-max-height: auto;
-}
-
-.cardContainer:hover {
-  transform: scale(1.05);
-}
-
-
-.imgContainer {
-  flex: 1;
-  border: none; 
-  background: none; 
-  padding: 0;
-  width: 50%; /* Fit image within the container */
-  height: 50%; /* Fit image within the container */
-  border-radius: 50%;
-  
-}
-
