--- src/components/CardsEliminar/Cards.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-import style from "./cards.module.css"
-import Card from "../Card/Card";
-function Cards () {
-  return (
-    <div className={style.container}>
-      <Card />
-    </div>
-  );
-};
-
-export default Cards;
