--- src/components/CardsEliminar/cards.module.css
@@ -1,6 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-    
-}
