--- src/components/CreateProduct/products.module.css
@@ -1,72 +1,0 @@
-.mainContainer {
-  background-color: rgba(53, 52, 50, 0.856);
-  padding: 50px;
-}
-
-.cont {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  background-color: rgba(201, 199, 197, 0.856);
-  margin: 10px;
-  border-radius: 5px;
-}
-
-.formContainer {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  background-color: rgba(201, 199, 197, 0.856);
-  margin: 10px;
-  padding: 10px;
-  border-radius: 10px;
-}
-
-.imgContainer {
-  flex: 1;
-  margin-right: 20px; /* Space between image and inputs */
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.imageCont {
-  border-radius: 50%;
-  background-color: rgb(255, 254, 254);
-  width: 150px; /* Adjust size as needed */
-  height: 150px; /* Adjust size as needed */
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.imgContainer img {
-  border-radius: 50%; /* Ensure the image itself is round */
-  width: 100%; /* Fit image within the container */
-  height: 80%; /* Fit image within the container */
-  object-fit: cover; /* Ensure image covers container while maintaining aspect ratio */
-}
-
-.inputsContainer {
-  flex: 2;
-}
-
-.inputRow {
-  display: flex;
-  gap: 10px; /* Space between input fields */
-  margin-bottom: 10px; /* Space between rows */
-}
-
-.inputsContainer input {
-  flex: 1;
-  padding: 8px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-}
-
-.btnContent {
-  display: flex;
-  flex-direction: row;
-  justify-content: end;
-  margin: 10px;
-}
