+++ src/components/ProductsConfig/ProductsConfig.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 import { useNavigate } from "react-router-dom";
 import ProductsConfigAdmin from "./ProductsConfigAdmin";
 import ProductsConfigSocio from "./ProductsConfigSocio";
