--- src/components/ProductsConfig/productsConfig.module.css
@@ -1,35 +1,0 @@
-.userList {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 10px; /* Espacio entre los bloques de usuario */
-}
-
-.container {
-  display: grid;
-  grid-template-columns: repeat(7, 1fr);
-  gap: 10px;
-  align-items: center;
-  padding: 10px;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  background-color: #f9f9f9;
-  width: 100%;
-  box-sizing: border-box;
-  margin: 0;
-}
-
-.content img {
-  width: 50px;
-  height: 50px;
-  border-radius: 50%;
-  object-fit: cover;
-}
-
-button {
-  cursor: pointer;
-  margin-right: 10px;
-}
-.container h2{
-
-  margin: 0;
-}
