--- src/components/UsersConfig/usersConfig.module.css
@@ -1,34 +1,0 @@
-.userList {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 10px; /* Espacio entre los bloques de usuario */
-}
-
-.container {
-  display: grid;
-  grid-template-columns: 190px 190px 190px 250px 190px 190px 190px;
-  gap: 10px;
-  align-items: center;
-  padding: 10px;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  background-color: #f9f9f9;
-  width: 100%;
-  box-sizing: border-box;
-}
-
-.content img {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 50px;
-  height: 50px;
-  border-radius: 50%;
-  object-fit: cover;
-}
-
-button {
-  cursor: pointer;
-  margin-right: 10px;
-}
-
