+++ src/index.js
@@ -10,7 +10,7 @@ import { GoogleOAuthProvider } from "@react-oauth/google";
 import axios from 'axios';
 
 
-// axios.defaults.baseURL = 'http://localhost:3001';
+axios.defaults.baseURL = 'http://localhost:3001';
 
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
@@ -18,8 +18,7 @@ root.render(
   <GoogleOAuthProvider clientId="29682485940-7n1d9hgf8pdicimp1vkqn4j99jus0nr4.apps.googleusercontent.com">
     <React.StrictMode>
       <Provider store={store}>
-        {" "}
-        {}
+      
         <App />
       </Provider>
     </React.StrictMode>
